run precommit in ci
