+++ .github/workflows/validate_csvs.yml
@@ -6,9 +6,13 @@ jobs:
   validate-csvs:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: run duplicate test
-        run: |
-          python3 tests/validate_csv.py
-          python3 tests/validate_bulk.py
+      - uses: actions/checkout@v6
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          enable-cache: true
+      - name: uv sync dependencies
+        working-directory: ./checks
+        run: uv sync --all-extra
+      - name: Run prek
+        uses: j178/prek-action@v2
