+++ .pre-commit-config.yaml
@@ -1,0 +1,42 @@
+# Copyright 2025 Lincoln Institute of Land Policy
+# SPDX-License-Identifier: Apache-2.0
+
+minimum_pre_commit_version: "2.9.0"
+ci:
+  autoupdate_schedule: monthly
+  skip: []
+repos:
+  - repo: meta
+    hooks:
+      - id: check-useless-excludes
+  - repo: local
+    hooks:
+      - id: forbid-symlinks
+        name: Forbid symlinks
+        entry: Forbid symlinks
+        language: fail
+        types: [symlink]
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v6.0.0
+    hooks:
+      - id: check-case-conflict
+      - id: check-executables-have-shebangs
+      - id: check-merge-conflict
+      - id: check-shebang-scripts-are-executable
+      - id: destroyed-symlinks
+      - id: detect-private-key
+      - id: fix-byte-order-marker
+      - id: mixed-line-ending
+  - repo: local
+    hooks: 
+      - id: check_metadata_against_jsonschema
+        name: Check Geoconnex metadata against jsonschema
+        entry: bash -c "cd ./checks && uv run --with sourcemeta-jsonschema jsonschema validate ./geoconnex_metadata_schema.json ../namespaces/"
+        language: system
+        pass_filenames: false
+
+      - id: validate_csv_and_metadata
+        name: validate info in csvs and metadata
+        entry: uv run checks/validate_csv_and_metadata.py
+        language: system
+        pass_filenames: false
