+++ .vscode/settings.json
@@ -4,7 +4,7 @@
             "fileMatch": [
                 "namespaces/**/metadata.json"
             ],
-            "url": "./.vscode/geoconnex_metadata_schema.json"
+            "url": "./checks/geoconnex_metadata_schema.json"
         }
     ]
 }
