+++ checks/README.md
@@ -1,0 +1,3 @@
+# Geoconnex Metadata Checks
+
+This directory contains scripts that can be used for checking the content of metadata in Geoconnex. For instance, whether sitemaps associated contact emails and have properly formatted redirect CSVs.
